Add files via upload
Completely deleted the returned table 
Changed delete and create library tables to reflect that change
Created triggers for the database
when an item is borrowed, the stock of that item will go down by one
when an item is returned the stock of that item will go up by one
If the date an item is returned is past its due by date, they will be assigned a fee. 
If no due back by date is supplied, one will be added 3 weeks for the current day.